+++ src/main/java/dev/muon/medievalorigins/client/MedievalOriginsClient.java
@@ -5,20 +5,15 @@ import dev.muon.medievalorigins.client.render.SummonedSkeletonRenderer;
 import dev.muon.medievalorigins.client.render.SummonedWitherSkeletonRenderer;
 import dev.muon.medievalorigins.client.render.SummonedZombieRenderer;
 import dev.muon.medievalorigins.entity.ModEntities;
-import dev.muon.medievalorigins.util.PowerCache;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientEntityEvents;
-import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
-import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
 import net.minecraft.client.KeyMapping;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.lwjgl.glfw.GLFW;
-
 
 @Environment(EnvType.CLIENT)
 public class MedievalOriginsClient implements ClientModInitializer {
@@ -37,13 +32,5 @@ public class MedievalOriginsClient implements ClientModInitializer {
         EntityRendererRegistry.register(ModEntities.SUMMON_ZOMBIE, SummonedZombieRenderer::new);
         EntityRendererRegistry.register(ModEntities.SUMMON_WITHER_SKELETON, SummonedWitherSkeletonRenderer::new);
         EntityRendererRegistry.register(ModEntities.SUMMONED_ARROW, SummonedArrowRenderer::new);
-        
-        initClientPowerCache();
-    }
-    
-    private static void initClientPowerCache() {
-        ClientTickEvents.END_CLIENT_TICK.register(client -> PowerCache.tick());
-        ClientEntityEvents.ENTITY_UNLOAD.register((entity, world) -> PowerCache.invalidate(entity));
-        ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> PowerCache.clearAll());
     }
 }