+++ src/main/java/dev/muon/medievalorigins/mixin/compat/icarus/IcarusHelperMixin.java
@@ -7,7 +7,7 @@ import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import dev.cammiescorner.icarus.api.IcarusPlayerValues;
 import dev.cammiescorner.icarus.util.IcarusHelper;
 import dev.muon.medievalorigins.power.IcarusWingsPowerType;
-import dev.muon.medievalorigins.util.PowerCache;
+import io.github.apace100.apoli.component.PowerHolderComponent;
 import io.github.apace100.apoli.power.type.PowerType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
@@ -23,7 +23,7 @@ public abstract class IcarusHelperMixin {
 
     @ModifyReturnValue(method = "getConfigValues", at = @At("RETURN"))
     private static IcarusPlayerValues modifyConfigValues(IcarusPlayerValues original, LivingEntity entity) {
-        var icarusWings = PowerCache.getFirstPowerType(entity, IcarusWingsPowerType.class, PowerType::isActive);
+        var icarusWings = PowerHolderComponent.getPowerTypes(entity, IcarusWingsPowerType.class).stream().filter(PowerType::isActive).findFirst();
         if (icarusWings.isEmpty()) return original;
 
         //TODO: Add as fields to power json
@@ -69,6 +69,26 @@ public abstract class IcarusHelperMixin {
             }
 
             @Override
+            public boolean dropOutOfSkyWhenTired() {
+                return false;
+            }
+
+            @Override
+            public boolean useStaminaForFlight() {
+                return false;
+            }
+
+            @Override
+            public float staminaAmount() {
+                return 0;
+            }
+
+            @Override
+            public float staminaRegen() {
+                return 0;
+            }
+
+            @Override
             public float requiredFoodAmount() {
                 return 0;
             }
@@ -77,13 +97,13 @@ public abstract class IcarusHelperMixin {
 
     @WrapOperation(method = "hasWings", at = @At(value = "INVOKE", target = "Ljava/util/function/Predicate;test(Ljava/lang/Object;)Z"))
     private static boolean hasWingsFromOrigin(Predicate<LivingEntity> instance, Object entity, Operation<Boolean> original) {
-        return PowerCache.hasPowerType((LivingEntity) entity, IcarusWingsPowerType.class)
+        return PowerHolderComponent.hasPowerType((LivingEntity) entity, IcarusWingsPowerType.class)
                 || original.call(instance, entity);
     }
 
     @WrapOperation(method = "getEquippedWings", at = @At(value = "INVOKE", target = "Ljava/util/function/Function;apply(Ljava/lang/Object;)Ljava/lang/Object;"))
     private static Object getOriginWings(Function<LivingEntity, ItemStack> instance, Object entity, Operation<ItemStack> original) {
-        if (PowerCache.hasPowerType((LivingEntity) entity, IcarusWingsPowerType.class)) {
+        if (PowerHolderComponent.hasPowerType((LivingEntity) entity, IcarusWingsPowerType.class)) {
             return null;
         }
         return original.call(instance, entity);